initialised some values

+++ ShooterGame/Scenes/GameScene.cpp
@@ -74,7 +74,7 @@ void GameScene::OnKeyboard(int key, bool down)
 	Camera* mCamera = Game::GetGame()->GetGameCamera();
 	float moveSensitivity = 32.0f;
 
-	for (size_t i = 0; i < inputManager->GetKeysSize(); i++)
+	for (int i = 0; i < inputManager->GetKeysSize(); i++)
 	{
 		Command* command = nullptr;
 		if (inputManager->GetKey(i))
@@ -87,8 +87,7 @@ void GameScene::OnKeyboard(int key, bool down)
 
 	}
 
-
-
+	// TODO replace these with commands
 	if (inputManager->GetKey(GLFW_KEY_R))
 	{
 		AddEntity(new Enemy());
